# docs/api/backend.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShopSteward Backend API — 设计契约（尚未实现）",
    "version": "0.2.0-draft",
    "description": "B0独立业务后端；B2为未来接口。Swagger展示不代表服务已实现。业务语义见../backend-development.md。"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "规划的本地backend地址，当前未启动"
    }
  ],
...
  "paths": {
    "/api/v1/catalog": {
      "get": {
        "operationId": "get_catalog",
        "tags": [
          "Dashboard"
        ],
        "summary": "查询商品和供应商报价",
        "description": "查询商品和供应商报价",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "viewer"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                },
                "example": {
                  "store_id": "store_001",
                  "currency": "CNY",
                  "products": [
                    {
                      "sku_id": "sku_001",
                      "name": "示例商品",
                      "unit": "piece"
                    }
                  ],
                  "offers": [
                    {
                      "supplier_id": "supplier_001",
                      "sku_id": "sku_001",
                      "unit_price_minor": 1000,
                      "minimum_order_quantity": 20,
                      "pack_size": 20,
                      "offer_version": "offer-1",
                      "currency": "CNY",
                      "lead_time_seconds": 86400,
                      "valid_from": "2026-09-06T00:00:00Z",
                      "valid_until": "2026-09-07T00:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/v1/events/batches": {
      "post": {
        "operationId": "ingest_event_batch",
        "tags": [
          "Internal"
        ],
        "summary": "接收并原子应用源事件批次",
        "description": "与主动拉取共用去重。全批校验；sequence缺口或重复ID不同内容返回409。",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "simulation:events"
        ],
        "security": [
          {
            "ServiceBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "同一逻辑命令网络重试复用；同键不同内容返回409。",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventBatchResult"
                },
                "example": {
                  "accepted_event_ids": [
                    "event_001"
                  ],
                  "duplicate_event_ids": [],
                  "last_sequence": 1
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本、幂等或业务并发冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatch"
              },
              "example": {
                "source": "simulation",
                "scenario_run_id": "scenario_001",
                "events": [
                  {
                    "event_id": "event_001",
                    "sequence": 1,
                    "schema_version": "1.0",
                    "event_type": "SALE_RECORDED",
                    "store_id": "store_001",
                    "occurred_at": "2026-09-06T00:00:00Z",
                    "simulation_time": "2026-09-06T00:00:00Z",
                    "payload": {
                      "sku_id": "sku_001",
                      "quantity": 10,
                      "unit_price_minor": 2000
                    }
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/dev/v1/scenarios": {
      "post": {
        "operationId": "create_dev_scenario",
        "tags": [
          "Development"
        ],
        "summary": "新建隔离的模拟运行及门店",
        "description": "仅DEV_API_ENABLED开放；异步初始化。scenario_run_id/store_id可在受理时为空，完成后从JobRun结果references取得。",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "admin"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "同一逻辑命令网络重试复用；同键不同内容返回409。",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "202": {
            "description": "已持久化受理，不表示业务执行完成",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioAccepted"
                },
                "example": {
                  "job_run_id": "job_init_001",
                  "status": "READY",
                  "scenario_run_id": null,
                  "store_id": null
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本、幂等或业务并发冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreate"
              },
              "example": {
                "scenario": "SC01"
              }
            }
          }
        }
      }
    },
    "/dev/v1/scenarios/{run_id}/advance": {
      "post": {
        "operationId": "advance_dev_scenario",
        "tags": [
          "Development"
        ],
        "summary": "显式推进模拟场景",
        "description": "显式推进模拟场景",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "admin"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "同一逻辑命令网络重试复用；同键不同内容返回409。",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "202": {
            "description": "已持久化受理，不表示业务执行完成",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                },
                "example": {
                  "job_run_id": "job_001",
                  "status": "READY",
                  "merged": false
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本、幂等或业务并发冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioAdvance"
              },
              "example": {
                "steps": 1
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "用户身份；角色与门店范围由服务端验证。"
      },
      "ServiceBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "服务身份；Agent无审批权限，simulation仅提交授权事件。"
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "retryable": {
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable",
          "details"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Offer": {
        "type": "object",
        "properties": {
          "supplier_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "sku_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "unit_price_minor": {
            "type": "integer",
            "minimum": 1
          },
          "minimum_order_quantity": {
            "type": "integer",
            "minimum": 1
          },
          "pack_size": {
            "type": "integer",
            "minimum": 1
          },
          "offer_version": {
            "type": "string",
            "minLength": 1
          },
          "currency": {
            "type": "string",
            "enum": [
              "CNY"
            ]
          },
          "lead_time_seconds": {
            "type": "integer",
            "minimum": 0
          },
          "valid_from": {
            "type": "string",
            "format": "date-time"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "supplier_id",
          "sku_id",
          "unit_price_minor",
          "minimum_order_quantity",
          "pack_size",
          "offer_version",
          "currency",
          "lead_time_seconds",
          "valid_from",
          "valid_until"
        ],
        "additionalProperties": false
      },
      "Product": {
        "type": "object",
        "properties": {
          "sku_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string",
            "enum": [
              "piece"
            ]
          }
        },
        "required": [
          "sku_id",
          "name",
          "unit"
        ],
        "additionalProperties": false
      },
      "Catalog": {
        "type": "object",
        "properties": {
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "currency": {
            "type": "string",
            "enum": [
              "CNY"
            ]
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "offers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Offer"
            }
          }
        },
        "required": [
          "store_id",
          "currency",
          "products",
          "offers"
        ],
        "additionalProperties": false
      },
...
      "JobAccepted": {
        "type": "object",
        "properties": {
          "job_run_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "RUNNING",
              "RETRY_WAIT"
            ]
          },
          "merged": {
            "type": "boolean"
          }
        },
        "required": [
          "job_run_id",
          "status",
          "merged"
        ],
        "additionalProperties": false
      },
...
      "ScenarioCreate": {
        "type": "object",
        "properties": {
          "scenario": {
            "type": "string",
            "enum": [
              "SC01"
            ]
          }
        },
        "required": [
          "scenario"
        ],
        "additionalProperties": false
      },
      "ScenarioAdvance": {
        "type": "object",
        "properties": {
          "steps": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": [
          "steps"
        ],
        "additionalProperties": false
      },
      "ScenarioAccepted": {
        "type": "object",
        "properties": {
          "job_run_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "RUNNING",
              "RETRY_WAIT"
            ]
          },
          "scenario_run_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "store_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "job_run_id",
          "status",
          "scenario_run_id",
          "store_id"
        ],
        "additionalProperties": false
      },
...
      "SaleEvent": {
        "type": "object",
        "properties": {
          "event_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "sequence": {
            "type": "integer",
            "minimum": 1
          },
          "schema_version": {
            "type": "string",
            "enum": [
              "1.0"
            ]
          },
          "event_type": {
            "type": "string",
            "enum": [
              "SALE_RECORDED"
            ]
          },
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "occurred_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "simulation_time": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "payload": {
            "$ref": "#/components/schemas/SalePayload"
          }
        },
        "required": [
          "event_id",
          "sequence",
          "schema_version",
          "event_type",
          "store_id",
          "occurred_at",
          "simulation_time",
          "payload"
        ],
        "additionalProperties": false
      },
      "ReceivedEvent": {
        "type": "object",
        "properties": {
          "event_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "sequence": {
            "type": "integer",
            "minimum": 1
          },
          "schema_version": {
            "type": "string",
            "enum": [
              "1.0"
            ]
          },
          "event_type": {
            "type": "string",
            "enum": [
              "GOODS_RECEIVED"
            ]
          },
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "occurred_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "simulation_time": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "payload": {
            "$ref": "#/components/schemas/ReceivedPayload"
          }
        },
        "required": [
          "event_id",
          "sequence",
          "schema_version",
          "event_type",
          "store_id",
          "occurred_at",
          "simulation_time",
          "payload"
        ],
        "additionalProperties": false
      },
      "DemandEvent": {
        "type": "object",
        "properties": {
          "event_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "sequence": {
            "type": "integer",
            "minimum": 1
          },
          "schema_version": {
            "type": "string",
            "enum": [
              "1.0"
            ]
          },
          "event_type": {
            "type": "string",
            "enum": [
              "DEMAND_REVISED"
            ]
          },
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "occurred_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "simulation_time": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "payload": {
            "$ref": "#/components/schemas/DemandPayload"
          }
        },
        "required": [
          "event_id",
          "sequence",
          "schema_version",
          "event_type",
          "store_id",
          "occurred_at",
          "simulation_time",
          "payload"
        ],
        "additionalProperties": false
      },
      "PurchaseEvent": {
        "type": "object",
        "properties": {
          "event_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "sequence": {
            "type": "integer",
            "minimum": 1
          },
          "schema_version": {
            "type": "string",
            "enum": [
              "1.0"
            ]
          },
          "event_type": {
            "type": "string",
            "enum": [
              "PURCHASE_ACCEPTED"
            ]
          },
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "occurred_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "simulation_time": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "payload": {
            "$ref": "#/components/schemas/PurchasePayload"
          }
        },
        "required": [
          "event_id",
          "sequence",
          "schema_version",
          "event_type",
          "store_id",
          "occurred_at",
          "simulation_time",
          "payload"
        ],
        "additionalProperties": false
      },
      "BusinessEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/SaleEvent"
          },
          {
            "$ref": "#/components/schemas/ReceivedEvent"
          },
          {
            "$ref": "#/components/schemas/DemandEvent"
          },
          {
            "$ref": "#/components/schemas/PurchaseEvent"
          }
        ],
        "discriminator": {
          "propertyName": "event_type",
          "mapping": {
            "SALE_RECORDED": "#/components/schemas/SaleEvent",
            "GOODS_RECEIVED": "#/components/schemas/ReceivedEvent",
            "DEMAND_REVISED": "#/components/schemas/DemandEvent",
            "PURCHASE_ACCEPTED": "#/components/schemas/PurchaseEvent"
          }
        }
      },
      "EventBatch": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "simulation"
            ]
          },
          "scenario_run_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessEvent"
            },
            "minItems": 1,
            "maxItems": 100
          }
        },
        "required": [
          "source",
          "scenario_run_id",
          "events"
        ],
        "additionalProperties": false
      },
      "EventBatchResult": {
        "type": "object",
        "properties": {
          "accepted_event_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          "duplicate_event_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          "last_sequence": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "accepted_event_ids",
          "duplicate_event_ids",
          "last_sequence"
        ],
        "additionalProperties": false
      },
...
    }
  }
}